Moved Repair...WizardPage Improved ConnectionsToStructCommand
